Fix bug with computing big numbers

+++ src/main/scala/calculator/Expr.scala
@@ -1,7 +1,7 @@
 package calculator
 
 abstract class Expr
-case class Number(value: Double) extends Expr
+case class Number(value: BigDecimal) extends Expr
 case class UnaryOp(operator: String, arg: Expr) extends Expr
 case class BinaryOp(operator: String, left: Expr, right: Expr) extends Expr
